Moved _version.py to smsdk folder
