+++ smsdk/_version.py
